LayoutTests/fast/forms/submit-form-attributes-expected.txt

Tests the behavior of .formaction, .formenctype, .formmethod and .formtarget of HTMLInputElement and HTMLButtonElement.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


Ordinary values for input:
PASS input.formAction is ""
PASS input.formEnctype is ""
PASS input.formMethod is ""
PASS input.formTarget is ""
PASS input.formAction is "http://localhost/"
PASS input.formAction is "http://localhost/"
PASS input.formEnctype is "text/plain"
PASS input.formEnctype is "na"
PASS input.formMethod is "GET"
PASS input.formMethod is "ni"
PASS input.formTarget is "_blank"
PASS input.formTarget is "nu"
PASS input.formAction is "http://example.com/"
PASS input.formAction is "http://example.com/"
PASS input.formEnctype is "text/plain"
PASS input.formEnctype is "nota"
PASS input.formMethod is "POST"
PASS input.formMethod is "neta"
PASS input.formTarget is "http://example.com"
PASS input.formTarget is "nta"
Setting null for input:
PASS input.formEnctype is ""
PASS input.getAttribute("formEnctype") is null
PASS input.formEnctype is "null"
PASS input.formMethod is ""
PASS input.getAttribute("formMethod") is null
PASS input.formMethod is "null"
PASS input.formTarget is ""
PASS input.getAttribute("formTarget") is null
PASS input.formTarget is "null"
Setting undefined for input:
PASS input.formEnctype is "undefined"
PASS input.getAttribute("formEnctype") is "undefined"
PASS input.formEnctype is "undefined"
PASS input.formMethod is "undefined"
PASS input.getAttribute("formMethod") is "undefined"
PASS input.formMethod is "undefined"
PASS input.formTarget is "undefined"
PASS input.getAttribute("formTarget") is "undefined"
PASS input.formTarget is "undefined"
Setting non-string for input:
PASS input.formEnctype is "256"
PASS input.getAttribute("formEnctype") is "256"
PASS input.formEnctype is "256"
PASS input.formMethod is "256"
PASS input.getAttribute("formMethod") is "256"
PASS input.formMethod is "256"
PASS input.formTarget is "256"
PASS input.getAttribute("formTarget") is "256"
PASS input.formTarget is "256"
Ordinary values for button:
PASS button.formAction is ""
PASS button.formEnctype is ""
PASS button.formMethod is ""
PASS button.formTarget is ""
PASS button.formAction is "http://localhost/"
PASS button.formAction is "http://localhost/"
PASS button.formEnctype is "text/plain"
PASS button.formEnctype is "na"
PASS button.formMethod is "GET"
PASS button.formMethod is "na"
PASS button.formTarget is "_blank"
PASS button.formTarget is "na"
PASS button.formAction is "http://example.com/"
PASS button.formAction is "http://example.com/"
PASS button.formEnctype is "text/plain"
PASS button.formEnctype is "nota"
PASS button.formMethod is "POST"
PASS button.formMethod is "nota"
PASS button.formTarget is "http://example.com"
PASS button.formTarget is "nota"
Setting null for button:
PASS button.formEnctype is ""
PASS button.getAttribute("formEnctype") is null
PASS button.formEnctype is "null"
PASS button.formMethod is ""
PASS button.getAttribute("formMethod") is null
PASS button.formMethod is "null"
PASS button.formTarget is ""
PASS button.getAttribute("formTarget") is null
PASS button.formTarget is "null"
Setting undefined for button:
PASS button.formEnctype is "undefined"
PASS button.getAttribute("formEnctype") is "undefined"
PASS button.formEnctype is "undefined"
PASS button.formMethod is "undefined"
PASS button.getAttribute("formMethod") is "undefined"
PASS button.formMethod is "undefined"
PASS button.formTarget is "undefined"
PASS button.getAttribute("formTarget") is "undefined"
PASS button.formTarget is "undefined"
Setting non-string for button:
PASS button.formEnctype is "256"
PASS button.getAttribute("formEnctype") is "256"
PASS button.formEnctype is "512"
PASS button.formMethod is "128"
PASS button.getAttribute("formMethod") is "128"
PASS button.formMethod is "17"
PASS button.formTarget is "100"
PASS button.getAttribute("formTarget") is "100"
PASS button.formTarget is "281"
PASS successfullyParsed is true

TEST COMPLETE
